% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_context.R
\name{compare_context}
\alias{compare_context}
\title{Internal function to compare contexts}
\usage{
compare_context(ctx, target, strict = TRUE)
}
\arguments{
\item{ctx, target}{context string}

\item{strict}{whether \code{ctx} equaling to \code{target} is
accepted}
}
\value{
Whether \code{ctx} is greater or equal than \code{target}
}
\description{
Internal function to compare contexts
}
