% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notch-diagnosis.R
\name{export_diagnose_voltage}
\alias{export_diagnose_voltage}
\title{Export diagnostic plot (\code{\link{pwelch}}) for 'RAVE' subjects}
\usage{
export_diagnose_voltage(
  subject,
  electrodes,
  blocks,
  h5_names = c("/notch/{block}", "/raw/{block}"),
  save_dir = "./export",
  width = 12,
  height = 7,
  useDingbats = FALSE,
  onefile = TRUE,
  winlens,
  freq_lims,
  nclass = 20,
  cex = 1,
  col = c(fore_col, back_col),
  ...,
  fore_col = "black",
  back_col = "grey80"
)
}
\arguments{
\item{subject}{character, or \code{\link[raveio]{RAVESubject}}}

\item{electrodes}{electrodes to generate plots}

\item{blocks}{the session block number}

\item{h5_names}{the 'HDF5' data names used}

\item{save_dir}{where to save the 'PDF' files}

\item{width, height, useDingbats, onefile, ...}{passed to \code{\link[grDevices]{pdf}}}

\item{winlens, freq_lims}{see \code{\link{pwelch}}}

\item{nclass}{number of bins in the histogram}

\item{cex}{font size}

\item{col, fore_col, back_col}{the color to use for different types of signals;
by default signals after the 'Notch' filters uses \code{fore_col} color
(black), and original trace uses \code{back_col} (grey) color}
}
\value{
Nothing, however, one or multiple 'PDF' files will be generated at
\code{save_dir} directory.
}
\description{
Export diagnostic plot (\code{\link{pwelch}}) for 'RAVE' subjects
}
\examples{

# Please download RAVE demo subject first
if(interactive()){
  export_path <- tempfile()
  export_diagnose_voltage("demo/DemoSubject", 13:16,
                          "008", save_dir = export_path)
  # open the following path
  normalizePath(export_path)
}

}
