% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-utils.R
\name{filter-window}
\alias{filter-window}
\alias{hanning}
\alias{hamming}
\alias{blackman}
\alias{blackmannuttall}
\alias{blackmanharris}
\alias{flattopwin}
\alias{bohmanwin}
\title{Filter window functions}
\usage{
hanning(n)

hamming(n)

blackman(n)

blackmannuttall(n)

blackmanharris(n)

flattopwin(n)

bohmanwin(n)
}
\arguments{
\item{n}{number of time-points in window}
}
\value{
A numeric vector of window with length \code{n}
}
\description{
Filter window functions
}
\examples{

hanning(10)
hamming(11)
blackmanharris(21)

}
