% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-diagnosis.R
\name{freqz2}
\alias{freqz2}
\title{Frequency response of digital filter}
\usage{
freqz2(b, a = 1, fs = 2 * pi, n = 512, whole = FALSE, ...)
}
\arguments{
\item{b}{the moving-average coefficients of an \code{ARMA} model}

\item{a}{the auto-regressive coefficients of an \code{ARMA} filter; default
is \code{1}}

\item{fs}{sampling frequency in \code{Hz}}

\item{n}{number of points at which to evaluate the frequency response;
default is \code{512}}

\item{whole}{whether to evaluate beyond \code{Nyquist} frequency; default
is false}

\item{...}{ignored}
}
\value{
A list of frequencies and corresponding responses in complex vector
}
\description{
Compute the z-plane frequency response of an \code{ARMA} model.
}
