% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{internal_rave_function}
\alias{internal_rave_function}
\title{Get external function from 'RAVE'}
\usage{
internal_rave_function(name, pkg, inherit = TRUE, on_missing = NULL)
}
\arguments{
\item{name}{function or variable name}

\item{pkg}{'RAVE' package name}

\item{inherit}{passed to \code{\link{get0}}}

\item{on_missing}{default value to return of no function is found}
}
\value{
Function object if found, otherwise \code{on_missing}.
}
\description{
Internal function used for examples relative to 'RAVE' project and should
not be used directly.
}
