% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{is_not_cran}
\alias{is_not_cran}
\title{Internal function}
\usage{
is_not_cran(if_interactive = TRUE, verbose = FALSE)
}
\arguments{
\item{if_interactive, verbose}{default is \code{TRUE}}
}
\value{
logical
}
\description{
Do not call this function directly
}
\keyword{internal}
