% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{left_hippocampus_mask}
\alias{left_hippocampus_mask}
\title{Left 'Hippocampus' of 'N27-Collin' brain}
\format{
A three-mode integer mask array with values of \code{1} ('Hippocampus')
and \code{0} (other brain tissues)
}
\usage{
left_hippocampus_mask
}
\description{
Left 'Hippocampus' of 'N27-Collin' brain
}
\keyword{datasets}
