% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Matrix4.R
\name{new_matrix4}
\alias{new_matrix4}
\title{Create a \code{Matrix4} instance for \code{'Affine'} transform}
\usage{
new_matrix4()
}
\value{
A \code{Matrix4} instance
}
\description{
Create a \code{Matrix4} instance for \code{'Affine'} transform
}
\seealso{
\code{\link{new_vector3}}, \code{\link{new_quaternion}}
}
