% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{parallel-options}
\alias{parallel-options}
\alias{detect_threads}
\alias{ravetools_threads}
\title{Set or get thread options}
\usage{
detect_threads()

ravetools_threads(n_threads = "auto", stack_size = "auto")
}
\arguments{
\item{n_threads}{number of threads to set}

\item{stack_size}{Stack size (in bytes) to use for worker threads. The
default used for \code{"auto"} is 2MB on 32-bit systems and 4MB on 64-bit
systems.}
}
\value{
\code{detect_threads} returns an integer of default threads that
is determined by the number of \code{CPU} cores; \code{ravetools_threads}
returns nothing.
}
\description{
Set or get thread options
}
\examples{

if(interactive()){
  detect_threads()
  ravetools_threads(n_threads = 2)
}

}
