% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-catmull.R
\name{print.ravetools_curve}
\alias{print.ravetools_curve}
\title{Print method for \code{ravetools_curve}}
\usage{
\method{print}{ravetools_curve}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ravetools_curve}.}

\item{...}{additional arguments passed to \code{\link{print.data.frame}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a concise summary of a \code{ravetools_curve} object
returned by \code{\link{catmull_rom_3d}}.
}
