% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwelch.R
\name{pwelch}
\alias{pwelch}
\alias{print.pwelch}
\alias{plot.pwelch}
\title{Plot 'Welch' Periodogram}
\usage{
pwelch(
  x,
  fs,
  window = 64,
  noverlap = 8,
  nfft = 256,
  col = "black",
  xlim = NULL,
  ylim = NULL,
  main = "Welch periodogram",
  plot = TRUE,
  log = "xy",
  cex = 1,
  ...
)

\method{print}{pwelch}(x, ...)

\method{plot}{pwelch}(
  x,
  log = c("xy", "x", "y", ""),
  type = "l",
  add = FALSE,
  col = 1,
  cex = 1,
  cex.main = cex,
  cex.sub = cex,
  cex.lab = cex * 0.8,
  cex.axis = cex * 0.7,
  las = 1,
  main = "Welch periodogram",
  xlab,
  ylab,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{\code{'pwelch'} instance returned by \code{pwelch} function}

\item{fs}{sample rate}

\item{window}{window length, default size is \code{64}}

\item{noverlap}{overlap between two adjacent windows, by default is \code{8}}

\item{nfft}{number of basis functions}

\item{col, xlim, ylim, main, type, cex, cex.main, cex.sub, cex.lab, cex.axis, las, xlab, ylab}{parameters passed to \code{\link[graphics]{plot.default}}}

\item{plot}{integer, whether to plot the result or not; choices are \code{0}, no plot; \code{1} plot on a new canvas; \code{2} add to existing canvas}

\item{log}{indicates which axis should be \code{\link{log10}} transformed; options are \code{''}, \code{'x'}, \code{'y'}, \code{'xy'}.}

\item{...}{will be passed to \code{plot.pwelch} or ignored}

\item{add}{logical, whether the plot should be added to existing canvas}
}
\description{
Plot 'Welch' Periodogram
}
