% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_context.R
\name{rave-context}
\alias{rave-context}
\alias{rave_context}
\alias{from_rave_context}
\title{Set or Get RAVE context}
\usage{
rave_context(context, package = NULL, module_id = NULL, instance = NULL)

from_rave_context(info = c("context", "package", "module_id", "instance"))
}
\arguments{
\item{context}{character or missing. If not missing, will set context
before returning current context}

\item{package}{package name, for example, \code{"ravebuiltins"},
this is required for \code{rave_module_debug} and above}

\item{module_id, instance}{RAVE module ID and instance object,
required for context \code{rave_running} and \code{rave_running_local}.}

\item{info}{which context information to obtain}
}
\value{
\code{rave_context} returns a \code{rave-context} object,
i.e. a list. \code{from_rave_context} returns corresponding
context information.
}
\description{
Set or Get RAVE context
}
\details{
There are four contexts in RAVE: \code{default},
\code{rave_module_debug}, \code{rave_running}, and
\code{rave_running_local}. These four contexts are ordered, with
\code{default} being the lowest and \code{rave_running_local} the
highest. The order only has internal meaning and users don't have to
remember it. However, the four contexts correspond to different
scenarios.

When creating a 'S3' generic function (see
\code{\link{rave_context_generics}}), a \code{default}
function must be provided to handle the default action of the
function.

\code{rave_module_debug} is for module developers to use
as it minimizes many interactive procedures and dynamically load
modules to simplify debugging process. Normally this context is
only used in development mode.

\code{rave_running} means when start-RAVE command is called, and a
shiny app launches within the browser. Many functions might change
their behaviors here. For example, data loading process might need
to show a modal dialogue instead of asking in the console.

\code{rave_running_local} means when module is running without
shiny. In this case, and shiny interactive contents need to be
switched back to console-based. What differs
\code{rave_running_local} and \code{default} is in the prior
context, the module has more information about the state and
one can always assume the data has been loaded and use them without
validating them. Under \code{default} context, nothing can be
assumed.
}
\seealso{
\code{\link{rave_context_generics}}
}
