% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{rave-states}
\alias{rave-states}
\alias{shiny_is_running}
\alias{rave_is_running}
\title{Check States of Current RAVE instances or 'shiny' instances}
\usage{
shiny_is_running()

rave_is_running()
}
\value{
Logical, for \code{shiny_is_running}, the result
indicates whether shiny is running. For \code{rave_is_running},
the result shows whether rave has
context higher/equaling to \code{'rave_running'}.
}
\description{
Check States of Current RAVE instances or 'shiny' instances
}
