% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rave_glue}
\alias{rave_glue}
\title{Wrapper around \code{\link[glue]{glue}}}
\usage{
rave_glue(..., .envir = parent.frame())
}
\arguments{
\item{..., .envir}{passed to \code{\link[glue]{glue}}}
}
\description{
Wrapper around \code{\link[glue]{glue}}
}
