% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{register_js}
\alias{register_js}
\title{Register JavaScript Libraries for RAVE-shiny applications}
\usage{
register_js(rmd = FALSE)
}
\arguments{
\item{rmd}{whether used in R-markdown files}
}
\value{
None
}
\description{
Register JavaScript Libraries for RAVE-shiny applications
}
