% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging-registration.R
\name{register_volume}
\alias{register_volume}
\title{Imaging registration using \code{'NiftyReg'}}
\usage{
register_volume(
  source,
  target,
  method = c("rigid", "affine", "nonlinear"),
  interpolation = c("cubic", "trilinear", "nearest"),
  threads = detect_threads(),
  symmetric = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{source}{source imaging data, or a \code{'nifti'} file path;
for example, 'CT'}

\item{target}{target imaging data to align to; for example, 'MRI'}

\item{method}{method of transformation, choices are \code{'rigid'},
\code{'affine'}, or \code{'nonlinear'}}

\item{interpolation}{how volumes should be interpolated, choices are
\code{'cubic'}, \code{'trilinear'}, or \code{'nearest'}}

\item{threads, symmetric, verbose, ...}{see \code{\link[RNiftyReg]{niftyreg}}}
}
\value{
See \code{\link[RNiftyReg]{niftyreg}}
}
\description{
Registers 'CT' to 'MRI', or 'MRI' to another 'MRI'
}
\examples{

if(interactive()) {

source <- system.file("extdata", "epi_t2.nii.gz", package="RNiftyReg")
target <- system.file("extdata", "flash_t1.nii.gz", package="RNiftyReg")
aligned <- register_volume(source, target, verbose = FALSE)

source_img <- imager::as.cimg(array(aligned$source[[1]], dim = c(dim(aligned$source[[1]]), 1)))
target_img <- imager::as.cimg(array(aligned$target, dim = c(dim(aligned$target), 1)))
aligned_img <- imager::as.cimg(array(aligned$image, dim = c(dim(aligned$image), 1)))
par(mfrow = c(2, 2), mar = c(0.1, 0.1, 3.1, 0.1))
plot(source_img, frame = 30, axes = FALSE, main = "Source image")
plot(target_img, frame = 64, axes = FALSE, main = "Target image")
plot(aligned_img, frame = 64, axes = FALSE, main = "Aligned image")

# bucket fill and calculate differences
aligned_img[is.nan(aligned_img) | aligned_img <= 1] <- 1
target_img[is.nan(target_img) | aligned_img <= 1] <- 1
plot(diff, frame = 64, axes = FALSE,  main = "Percentage Difference")

}

}
