% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{safe_alert}
\alias{safe_alert}
\title{Show Alert in Shiny apps}
\usage{
safe_alert(
  title,
  message,
  type = c("warning", "error", "success", "info"),
  onConfirm = NULL,
  inputId = "shinyalert",
  closeOnEsc = FALSE,
  closeOnClickOutside = FALSE,
  html = FALSE,
  showCancelButton = FALSE,
  confirmButtonText = "OK",
  session = shiny::getDefaultReactiveDomain(),
  fancy = TRUE
)
}
\arguments{
\item{title, message}{title and message of the alert}

\item{type}{options are \code{"warning"}, \code{"error"},
\code{"success"}, and \code{"info"}}

\item{onConfirm}{R function to call upon confirmation}

\item{inputId}{input ID}

\item{closeOnEsc, closeOnClickOutside}{whether to easy close the modal}

\item{html}{whether not to escape the message}

\item{showCancelButton}{whether to show dismiss button}

\item{confirmButtonText}{text of confirm button}

\item{session}{shiny session}

\item{fancy}{whether try to use fancy version}
}
\description{
Show Alert in Shiny apps
}
