% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{stimulation_signal}
\alias{stimulation_signal}
\title{Sample stimulation recording}
\format{
A list of one-second signal trace and sample rate (30000)
}
\usage{
stimulation_signal
}
\description{
Sample stimulation recording
}
\keyword{datasets}
