% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcg.R
\name{vcg_max_edge_length}
\alias{vcg_max_edge_length}
\title{Maximum edge length of a triangular mesh}
\usage{
vcg_max_edge_length(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class \code{'mesh3d'}.}
}
\value{
A single numeric value: the maximum edge length, in mesh units.
}
\description{
Returns the length of the longest edge in the mesh.
}
\examples{
if (is_not_cran()) {

  sphere <- vcg_sphere()
  vcg_max_edge_length(sphere)

}

}
\seealso{
\code{\link{vcg_average_edge_length}},
  \code{\link{vcg_subdivide_max_edge_length}}
}
