% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcg.R
\name{vcg_sphere}
\alias{vcg_sphere}
\title{Simple 3-dimensional sphere mesh}
\usage{
vcg_sphere(sub_division = 3L, normals = TRUE)
}
\arguments{
\item{sub_division}{density of vertex in the resulting mesh}

\item{normals}{whether the normal vectors should be calculated}
}
\value{
A \code{'mesh3d'} object
}
\description{
Simple 3-dimensional sphere mesh
}
\examples{

vcg_sphere()

}
