% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet-rave.R
\name{wavelet_rave_subject}
\alias{wavelet_rave_subject}
\title{High-performance 'Morlet'-wavellet on 'RAVE' subject}
\usage{
wavelet_rave_subject(
  subject,
  electrode,
  frequencies,
  kernel_cycles,
  downsample_rate = 100,
  demean = TRUE,
  signal_source = "/notch/{block}"
)
}
\arguments{
\item{subject}{character of subject ID or \code{\link[raveio]{RAVESubject}}
instance}

\item{electrode}{length of one, which electrode should be applied}

\item{frequencies, kernel_cycles, demean}{see \code{\link{morlet_wavelet}}}

\item{downsample_rate}{down-sample rate after wavelet}

\item{signal_source}{which 'HDF5' name should be used. Current default is
\code{"/notch/{block}"} for 'Notch' filtered signals}
}
\value{
Current time-stamp. The wavelet power and phase will be saved to
'RAVE' structure automatically
}
\description{
This function is intended to be called within builtin
pipelines. Please check \url{https://rave.wiki} to see details.
}
