% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Quaternion.R
\name{new_quaternion}
\alias{new_quaternion}
\alias{as_quaternion}
\title{Create a \code{Quaternion} instance to store '3D' rotation}
\usage{
new_quaternion(x = 0, y = 0, z = 0, w = 1)

as_quaternion(q)
}
\arguments{
\item{x, y, z, w}{numeric of length one}

\item{q}{R object to be converted to \code{Quaternion}}
}
\value{
A \code{Quaternion} instance
}
\description{
Create instances that mimic the \code{'three.js'} syntax.
}
\seealso{
\code{\link{new_vector3}}, \code{\link{new_matrix4}}
}
