% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcg.R
\name{vcg_mesh_volume}
\alias{vcg_mesh_volume}
\title{Compute volume for manifold meshes}
\usage{
vcg_mesh_volume(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class \code{'mesh3d'}}
}
\value{
numeric; the volume of the mesh
}
\description{
Compute volume for manifold meshes
}
\examples{


volume <- array(0, dim = c(8,8,8))
volume[4:5, 4:5, 4:5] <- 1
mesh <- mesh_from_volume(
  volume, verbose = FALSE, remesh = FALSE, smooth = FALSE)

vcg_mesh_volume(mesh)


}
