% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcg.R
\name{vcg_mesh_volume}
\alias{vcg_mesh_volume}
\title{Compute volume for manifold meshes}
\usage{
vcg_mesh_volume(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class \code{'mesh3d'}}
}
\value{
The numeric volume of the mesh
}
\description{
Compute volume for manifold meshes
}
\examples{

# Initial mesh
mesh <- vcg_sphere()

vcg_mesh_volume(mesh)

}
