% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcg.R
\name{vcg_update_normals}
\alias{vcg_update_normals}
\title{Update vertex normal}
\usage{
vcg_update_normals(
  mesh,
  weight = c("area", "angle"),
  pointcloud = c(10, 0),
  verbose = FALSE
)
}
\arguments{
\item{mesh}{triangular mesh or a point-cloud (matrix of 3 columns)}

\item{weight}{method to compute per-vertex normal vectors: \code{"area"}
weighted average of surrounding face 'normals', or \code{"angle"} weighted
vertex 'normals'.}

\item{pointcloud}{integer vector of length 2: containing optional
parameters for normal calculation of point clouds; the first entry
specifies the number of neighboring points to consider; the second
entry specifies the amount of smoothing iterations to be performed.}

\item{verbose}{whether to verbose the progress}
}
\description{
Update vertex normal
}
\examples{

volume <- array(0, dim = c(8,8,8))
volume[4:5, 4:5, 4:5] <- 1
mesh <- mesh_from_volume(volume, verbose = FALSE)
mesh$normals <- NULL
new_mesh <- vcg_update_normals(mesh, weight = "angle",
                               pointcloud = c(10, 10))

if(!is_not_cran()) {
  rgl_view({
    rgl_call("mfrow3d", 1, 2)
    rgl_call("shade3d", mesh, col = 2)

    rgl_call("next3d")
    rgl_call("shade3d", new_mesh, col = 2)
  })
}


}
